This is a re-submission of #493 which for some reason got closed and can't be re-opened...

## What do these changes do?

This PR adds an Aarch64 assembly implementation of Int8 bitpacking.

## How Has This Been Tested?

CI. Specific tests have been added for Int8 Aarch64 bitpacking.

## Benchmark Results

These benchmarks were taken with the LCE benchmarking tool on my Android phone (a OnePlus 6), running the QuickNet models (modified to use Int8 instead of float) single-threaded. I used `num_runs=250` and report the average latency and the standard deviation.

| Model         | Baseline (#492)   | PR            |
|---------------|---------------|---------------|
| QuickNet      | 17.18 +- 0.09 | 16.22 +- 0.05 |
| QuickNetLarge | 24.59 +- 0.10 | 23.21 +- 0.07 |
| QuickNetXL    | 43.27 +- 0.09 | 40.32 +- 0.06 |

With the Ruy profiler, I've observed that with these changes the `LceQuantize` op accounts for `0.57%` of the single-threaded Int8 QuickNet latency.

## Related issue number

N/A.